# Remove First and Last data members from Metrics structs