--- a/report/metrics.go
+++ b/report/metrics.go
@@ -48,10 +48,16 @@ func (m Metrics) Copy() Metrics {
 // Metric is a list of timeseries data with some metadata. Clients must use the
 // Add method to add values.  Metrics are immutable.
 type Metric struct {
-	Samples     []Sample
-	Min, Max    float64
-	First, Last time.Time
+	Samples  []Sample
+	Min, Max float64
 }
+
+// Following two functions are exported for testing only: make sure there are samples before calling.
+// First gives the timestamp of the first sample
+func (m Metric) First() time.Time { return m.Samples[0].Timestamp }
+
+// Last gives the timestamp of the last sample
+func (m Metric) Last() time.Time { return m.Samples[len(m.Samples)-1].Timestamp }
 
 // Sample is a single datapoint of a metric.
 type Sample struct {
@@ -65,8 +71,6 @@ func MakeSingletonMetric(t time.Time, v float64) Metric {
 		Samples: []Sample{{t, v}},
 		Min:     v,
 		Max:     v,
-		First:   t,
-		Last:    t,
 	}
 
 }
@@ -97,8 +101,6 @@ func MakeMetric(samples []Sample) Metric {
 		Samples: samples,
 		Min:     min,
 		Max:     max,
-		First:   samples[0].Timestamp,
-		Last:    samples[len(samples)-1].Timestamp,
 	}
 }
 
@@ -108,28 +110,12 @@ func (m Metric) WithMax(max float64) Metric {
 		Samples: m.Samples,
 		Max:     max,
 		Min:     m.Min,
-		First:   m.First,
-		Last:    m.Last,
 	}
 }
 
 // Len returns the number of samples in the metric.
 func (m Metric) Len() int {
 	return len(m.Samples)
-}
-
-func first(t1, t2 time.Time) time.Time {
-	if t2.IsZero() || (!t1.IsZero() && t1.Before(t2)) {
-		return t1
-	}
-	return t2
-}
-
-func last(t1, t2 time.Time) time.Time {
-	if t2.IsZero() || (!t1.IsZero() && t1.After(t2)) {
-		return t1
-	}
-	return t2
 }
 
 // Merge combines the two Metrics and returns a new result.
@@ -141,7 +127,7 @@ func (m Metric) Merge(other Metric) Metric {
 		return other
 	case len(other.Samples) == 0:
 		return m
-	case other.First.After(m.Last):
+	case other.First().After(m.Last()):
 		samplesOut := make([]Sample, len(m.Samples)+len(other.Samples))
 		copy(samplesOut, m.Samples)
 		copy(samplesOut[len(m.Samples):], other.Samples)
@@ -149,10 +135,8 @@ func (m Metric) Merge(other Metric) Metric {
 			Samples: samplesOut,
 			Max:     math.Max(m.Max, other.Max),
 			Min:     math.Min(m.Min, other.Min),
-			First:   m.First,
-			Last:    other.Last,
 		}
-	case m.First.After(other.Last):
+	case m.First().After(other.Last()):
 		samplesOut := make([]Sample, len(m.Samples)+len(other.Samples))
 		copy(samplesOut, other.Samples)
 		copy(samplesOut[len(other.Samples):], m.Samples)
@@ -160,8 +144,6 @@ func (m Metric) Merge(other Metric) Metric {
 			Samples: samplesOut,
 			Max:     math.Max(m.Max, other.Max),
 			Min:     math.Min(m.Min, other.Min),
-			First:   other.First,
-			Last:    m.Last,
 		}
 	}
 
@@ -194,8 +176,6 @@ func (m Metric) Merge(other Metric) Metric {
 		Samples: samplesOut,
 		Max:     math.Max(m.Max, other.Max),
 		Min:     math.Min(m.Min, other.Min),
-		First:   first(m.First, other.First),
-		Last:    last(m.Last, other.Last),
 	}
 }
 
@@ -214,8 +194,6 @@ type WireMetrics struct {
 	Samples []Sample `json:"samples,omitempty"`
 	Min     float64  `json:"min"`
 	Max     float64  `json:"max"`
-	First   string   `json:"first,omitempty"`
-	Last    string   `json:"last,omitempty"`
 	dummySelfer
 }
 
@@ -241,8 +219,6 @@ func (m Metric) ToIntermediate() WireMetrics {
 		Samples: m.Samples,
 		Max:     m.Max,
 		Min:     m.Min,
-		First:   renderTime(m.First),
-		Last:    renderTime(m.Last),
 	}
 }
 
@@ -253,8 +229,6 @@ func (m WireMetrics) FromIntermediate() Metric {
 		Samples: m.Samples,
 		Max:     m.Max,
 		Min:     m.Min,
-		First:   parseTime(m.First),
-		Last:    parseTime(m.Last),
 	}
 }