… report/metrics_internal_test.go